feat: use TransactionRef in the runtime API
