+++ pallets/transaction-storage/primitives/src/lib.rs
@@ -19,7 +19,27 @@
 
 extern crate alloc;
 
+use codec::{Decode, Encode, MaxEncodedLen};
+use scale_info::TypeInfo;
+
 pub mod cids;
 
 /// 32-byte hash of a stored blob of data.
 pub type ContentHash = [u8; 32];
+
+/// Identifies a previously-stored entry in the pallet's `Transactions` map.
+#[derive(
+	Clone,
+	PartialEq,
+	Eq,
+	Debug,
+	Encode,
+	Decode,
+	codec::DecodeWithMemTracking,
+	TypeInfo,
+	MaxEncodedLen,
+)]
+pub enum TransactionRef<BlockNumber> {
+	Position { block: BlockNumber, index: u32 },
+	ContentHash(ContentHash),
+}
